requirements.txt
discord.py
aiosqlite>=0.19.0
python-dotenv
psutil
pytz
pillow